requirements.txt
altair                    == 5.0.1
anyio                     == 4.2.0
argon2-cffi               == 21.3.0
argon2-cffi-bindings      == 21.2.0
asttokens                 == 2.4.1
async-lru                 == 2.0.4
attrs                     == 23.1.0
Babel                     == 2.11.0
beautifulsoup4            == 4.12.3
bleach                    == 4.1.0
blinker                   == 1.6.2
Bottleneck                == 1.3.7
branca                    == 0.6.0
Brotli                    == 1.0.9
cachetools                == 5.3.3
certifi                   == 2024.7.4
cffi                      == 1.16.0
charset-normalizer        == 2.0.4
click                     == 8.1.7
colorama                  == 0.4.6
comm                      == 0.2.2
debugpy                   == 1.8.2
decorator                 == 5.1.1
defusedxml                == 0.7.1
exceptiongroup            == 1.2.2
executing                 == 2.0.1
fastjsonschema            == 2.16.2
folium                    == 0.17.0
gitdb                     == 4.0.7
GitPython                 == 3.1.43
idna                      == 3.7
importlib_metadata        == 8.0.0
inflection                == 0.5.1
ipykernel                 == 6.29.5
ipython                   == 8.18.1
ipython-genutils          == 0.2.0
ipywidgets                == 7.8.1
jedi                      == 0.19.1
Jinja2                    == 3.1.4
json5                     == 0.9.6
jsonschema                == 4.19.2
jsonschema-specifications == 2023.12.1
jupyter_client            == 8.6.2
jupyter_core              == 5.7.2
jupyter-events            == 0.10.0
jupyter-lsp               == 2.2.0
jupyter_server            == 2.14.1
jupyter_server_terminals  == 0.4.4
jupyterlab                == 4.0.11
jupyterlab-pygments       == 0.1.2
jupyterlab_server         == 2.25.1
jupyterlab-widgets        == 1.0.0
markdown-it-py            == 2.2.0
MarkupSafe                == 2.1.3
matplotlib-inline         == 0.1.7
mdurl                     == 0.1.0
mistune                   == 2.0.4
mkl-fft                   == 1.3.8
mkl-random                == 1.2.4
mkl-service               == 2.4.0
nbclient                  == 0.8.0
nbconvert                 == 7.10.0
nbformat                  == 5.10.4
nest_asyncio              == 1.6.0
notebook                  == 7.0.8
notebook_shim             == 0.2.3
numexpr                   == 2.8.7
numpy                     == 1.26.4
overrides                 == 7.4.0
packaging                 == 23.2
pandas                    == 2.2.2
pandocfilters             == 1.5.0
parso                     == 0.8.4
pickleshare               == 0.7.5
pillow                    == 10.4.0
pip                       == 24.0
platformdirs              == 4.2.2
plotly                    == 5.22.0
prometheus-client         == 0.14.1
prompt_toolkit            == 3.0.47
protobuf                  == 4.25.3
psutil                    == 6.0.0
pure-eval                 == 0.2.2
pyarrow                   == 16.1.0
pycparser                 == 2.21
pydeck                    == 0.8.0
Pygments                  == 2.18.0
PySocks                   == 1.7.1
python-dateutil           == 2.9.0.post0
python-json-logger        == 2.0.7
pytz                      == 2024.1
pywin32                   == 306
pywinpty                  == 2.0.10
PyYAML                    == 6.0.1
pyzmq                     == 26.0.3
referencing               == 0.30.2
requests                  == 2.32.2
rfc3339-validator         == 0.1.4
rfc3986-validator         == 0.1.1
rich                      == 13.7.1
rpds-py                   == 0.10.6
Send2Trash                == 1.8.2
setuptools                == 69.5.1
six                       == 1.16.0
smmap                     == 4.0.0
sniffio                   == 1.3.0
soupsieve                 == 2.5
stack-data                == 0.6.2
streamlit                 == 1.32.0
tenacity                  == 8.2.2
terminado                 == 0.17.1
tinycss2                  == 1.2.1
toml                      == 0.10.2
tomli                     == 2.0.1
toolz                     == 0.12.0
tornado                   == 6.4.1
traitlets                 == 5.14.3
typing_extensions         == 4.12.2
tzdata                    == 2023.3
urllib3                   == 2.2.2
watchdog                  == 4.0.1
wcwidth                   == 0.2.13
webencodings              == 0.5.1
websocket-client          == 1.8.0
wheel                     == 0.43.0
widgetsnbextension        == 3.6.6
win-inet-pton             == 1.1.0
xyzservices               == 2024.6.0
zipp                      == 3.19.2